+++ app/Http/Controllers/Api/Psychologist/PsychologistProfileController.php
@@ -303,13 +303,22 @@ class PsychologistProfileController extends Controller
     public function deletePsychologist()
     {
         $userDeleted = User::where('id', auth()->user()->id)->first();
-        $userDeleted->userLocation->delete();
-        $userDeleted->userPersonalData->delete();
+        if($userDeleted->userLocation != null)
+        {
+            $userDeleted->userLocation->delete();
+        }
+        if($userDeleted->userPersonalData != null)
+        {
+            $userDeleted->userPersonalData->delete();
+        }
         $userType = (auth()->user()->profile_id == '3') ? 'patient' : 'psychologist';
         $appointments = Appointment::where($userType.'_user_id', auth()->user()->id)->get();
-        $appointments->each(function ($appointment){
-            $appointment->delete();
-        });
+        if($appointments != null)
+        {
+            $appointments->each(function ($appointment){
+                $appointment->delete();
+            });
+        }
         $userDeleted->delete();
     }
 }
